requirements.txt
vllm==0.13.0
huggingface_hub
transformers
tqdm
wandb
ipywidgets
bitsandbytes
accelerate
datasets
hydra-core==1.3.2
evaluate
pydantic
trl
peft
deepspeed
cachetools
scikit-learn
jinja2
colorama
timeout_decorator
sympy
multiprocess
fastapi
pydantic
uvicorn
PyYAML
llmcompressor
polars
omegaconf==2.3.0
antlr4-python3-runtime==4.9.3
math-verify
google-generativeai
psutil
liger-kernel==0.5.8
# flash-attn==2.7.4.post1